Implement dynamic time warping
Previously, gestures were compared using interpolation based on length
followed by direct comparision of corresponding points. This leaves
the problem where one part of a gesture may be drawn longer and then
the rest of the gesture compares to the "wrong" points. Dynamic time
warping seeks to alleviate this problem.